Change scripts based on feedback, see description for details [DHDO-469]
Removed archive_file and replaced it with replicate_file.
Archiving file is dangerous in the sense that we use an unlink for file removal.
If file is on tape already and we unlink we lose the reference, hence archiving should only be done with rules.

replicate_file replicates the file file from one resource to the other. Sort of triangle (circular) structure.
A file on a resource will always be rplicated to the "next in line" resource. Last resource will replicate to the first resource.

Changed get_file so that it will "download" the file to the current directory, with the same name and a epoch timestamp.

Only thing left is checksum validation. Will be other commit.